Refactoring whitespace_before_token rule. Updating documentation.
