+++ docs/attribute_declaration_rules.rst
@@ -31,6 +31,8 @@ attribute_declaration_101
 
 This rule checks for at least a single space before the colon.
 
+|configuring_whitespace_rules_link|
+
 **Violation**
 
 .. code-block:: vhdl
