+++ docs/attribute_specification_rules.rst
@@ -35,6 +35,8 @@ attribute_specification_101
 
 This rule checks for a single space before the **is** keyword.
 
+|configuring_whitespace_rules_link|
+
 **Violation**
 
 .. code-block:: vhdl
