+++ docs/comment_rules.rst
@@ -10,6 +10,8 @@ comment_004
 
 This rule checks for at least a single space before inline comments.
 
+|configuring_whitespace_rules_link|
+
 **Violation**
 
 .. code-block:: vhdl
