+++ docs/entity_specification_rules.rst
@@ -33,6 +33,8 @@ entity_specification_101
 
 This rule checks for at least a single space before the colon.
 
+|configuring_whitespace_rules_link|
+
 **Violation**
 
 .. code-block:: vhdl
