+++ docs/sequential_rules.rst
@@ -58,6 +58,8 @@ sequential_003
 
 This rule checks for at least a single space before the **<=** operator.
 
+|configuring_whitespace_rules_link|
+
 **Violation**
 
 .. code-block:: vhdl
