+++ docs/variable_assignment_rules.rst
@@ -62,6 +62,8 @@ variable_assignment_003
 
 This rule checks for at least a single space before the assignment.
 
+|configuring_whitespace_rules_link|
+
 **Violation**
 
 .. code-block:: vhdl
