+++ docs/variable_rules.rst
@@ -114,6 +114,8 @@ variable_006
 
 This rule checks for at least a single space before the colon.
 
+|configuring_whitespace_rules_link|
+
 **Violation**
 
 .. code-block:: vhdl
