+++ vsg/rules/__init__.py
@@ -34,7 +34,6 @@ from .insert_carriage_return_after_token_if_it_is_not_followed_by_a_comment_and_
 from .align_consecutive_lines_starting_with_a_comment_above_line_starting_with_token import align_consecutive_lines_starting_with_a_comment_above_line_starting_with_token
 
 from .align_consecutive_lines_after_line_starting_with_token_and_stopping_with_token import align_consecutive_lines_after_line_starting_with_token_and_stopping_with_token
-from .whitespace_before_token import whitespace_before_token
 from .whitespace_before_tokens_in_between_tokens import whitespace_before_tokens_in_between_tokens
 from .remove_blank_lines_above_line_starting_with_token import remove_blank_lines_above_line_starting_with_token
 from .blank_lines_between_token_pairs import blank_lines_between_token_pairs
