+++ vsg/rules/attribute_declaration/rule_101.py
@@ -1,15 +1,17 @@
 
 from vsg import token
 
-from vsg.rules import whitespace_before_token
+from vsg.rules.whitespace_before_token import Rule
 
 lTokens = []
 lTokens.append(token.attribute_declaration.colon)
 
 
-class rule_101(whitespace_before_token):
+class rule_101(Rule):
     '''
     This rule checks for at least a single space before the colon.
+
+    |configuring_whitespace_rules_link|
 
     **Violation**
 
@@ -24,5 +26,5 @@ class rule_101(whitespace_before_token):
        attribute max_delay : time;
     '''
     def __init__(self):
-        whitespace_before_token.__init__(self, 'attribute_declaration', '101', lTokens)
+        Rule.__init__(self, 'attribute_declaration', '101', lTokens)
         self.number_of_spaces = '>=1'
