+++ vsg/rules/case/rule_005.py
@@ -1,5 +1,5 @@
 
-from vsg.rules import whitespace_before_token as Rule
+from vsg.rules.whitespace_before_token import Rule
 
 from vsg.token import case_statement_alternative as token
 
@@ -7,6 +7,8 @@ from vsg.token import case_statement_alternative as token
 class rule_005(Rule):
     '''
     This rule checks for a single space before the **=>** operator.
+
+    |configuring_whitespace_rules_link|
 
     **Violation**
 
@@ -26,4 +28,3 @@ class rule_005(Rule):
     '''
     def __init__(self):
         Rule.__init__(self, 'case', '005', [token.assignment])
-        self.solution = 'Reduce spaces before the assignment operator to a single space.'
