+++ vsg/rules/comment/rule_004.py
@@ -1,15 +1,17 @@
 
 from vsg import parser
 
-from vsg.rules import whitespace_before_token
+from vsg.rules.whitespace_before_token import Rule
 
 lTokens = []
 lTokens.append(parser.comment)
 
 
-class rule_004(whitespace_before_token):
+class rule_004(Rule):
     '''
     This rule checks for at least a single space before inline comments.
+
+    |configuring_whitespace_rules_link|
 
     **Violation**
 
@@ -27,6 +29,5 @@ class rule_004(whitespace_before_token):
     '''
 
     def __init__(self):
-        whitespace_before_token.__init__(self, 'comment', '004', lTokens)
-        self.solution = 'Add a single space before comment'
+        Rule.__init__(self, 'comment', '004', lTokens)
         self.number_of_spaces = '>=1'
