+++ vsg/rules/constant/rule_010.py
@@ -1,16 +1,18 @@
 
 from vsg import token
 
-from vsg.rules import whitespace_before_token
+from vsg.rules.whitespace_before_token import Rule
 
 lTokens = []
 lTokens.append(token.constant_declaration.assignment_operator)
 
 
-class rule_010(whitespace_before_token):
+class rule_010(Rule):
     '''
     This rule checks for a single space before the := keyword in constant declarations.
     Having a space makes it clearer where the assignment occurs on the line.
+
+    |configuring_whitespace_rules_link|
 
     **Violation**
 
@@ -27,6 +29,5 @@ class rule_010(whitespace_before_token):
        constant width : integer := 10;
     '''
     def __init__(self):
-        whitespace_before_token.__init__(self, 'constant', '010', lTokens)
-        self.solution = 'Ensure at least one space before :=.'
+        Rule.__init__(self, 'constant', '010', lTokens)
         self.number_of_spaces = '>=1'
