+++ vsg/rules/entity_specification/rule_101.py
@@ -1,15 +1,17 @@
 
 from vsg import token
 
-from vsg.rules import whitespace_before_token
+from vsg.rules.whitespace_before_token import Rule
 
 lTokens = []
 lTokens.append(token.entity_specification.colon)
 
 
-class rule_101(whitespace_before_token):
+class rule_101(Rule):
     '''
     This rule checks for at least a single space before the colon.
+
+    |configuring_whitespace_rules_link|
 
     **Violation**
 
@@ -28,5 +30,5 @@ class rule_101(whitespace_before_token):
        attribute coordinate of comp_1     : component is (0.0, 17.5);
     '''
     def __init__(self):
-        whitespace_before_token.__init__(self, 'entity_specification', '101', lTokens)
+        Rule.__init__(self, 'entity_specification', '101', lTokens)
         self.number_of_spaces = '>=1'
