+++ vsg/rules/loop_statement/rule_102.py
@@ -1,7 +1,7 @@
 
 from vsg import token
 
-from vsg.rules import whitespace_before_token as Rule
+from vsg.rules.whitespace_before_token import Rule
 
 lTokens = []
 lTokens.append(token.loop_statement.loop_keyword)
@@ -10,6 +10,8 @@ lTokens.append(token.loop_statement.loop_keyword)
 class rule_102(Rule):
     '''
     This rule checks for a single space before the **loop** keyword.
+
+    |configuring_whitespace_rules_link|
 
     **Violation**
