+++ vsg/rules/sequential/rule_003.py
@@ -1,7 +1,7 @@
 
 from vsg import token
 
-from vsg.rules import whitespace_before_token
+from vsg.rules.whitespace_before_token import Rule
 
 lTokens = []
 lTokens.append(token.simple_force_assignment.assignment)
@@ -9,9 +9,11 @@ lTokens.append(token.simple_waveform_assignment.assignment)
 lTokens.append(token.simple_release_assignment.assignment)
 
 
-class rule_003(whitespace_before_token):
+class rule_003(Rule):
     '''
     This rule checks for at least a single space before the **<=** operator.
+
+    |configuring_whitespace_rules_link|
 
     **Violation**
 
@@ -28,6 +30,5 @@ class rule_003(whitespace_before_token):
        rd_en   <= '0';
     '''
     def __init__(self):
-        whitespace_before_token.__init__(self, 'sequential', '003', lTokens)
-        self.solution = 'Ensure at least a single space exists before the *<=* keyword.'
+        Rule.__init__(self, 'sequential', '003', lTokens)
         self.number_of_spaces = '>=1'
