+++ vsg/rules/variable_assignment/rule_003.py
@@ -1,7 +1,7 @@
 
 from vsg import token
 
-from vsg.rules import whitespace_before_token
+from vsg.rules.whitespace_before_token import Rule
 
 lTokens = []
 lTokens.append(token.simple_variable_assignment.assignment)
@@ -9,9 +9,11 @@ lTokens.append(token.conditional_variable_assignment.assignment)
 lTokens.append(token.selected_variable_assignment.assignment)
 
 
-class rule_003(whitespace_before_token):
+class rule_003(Rule):
     '''
     This rule checks for at least a single space before the assignment.
+
+    |configuring_whitespace_rules_link|
 
     **Violation**
 
@@ -28,6 +30,5 @@ class rule_003(whitespace_before_token):
          count := counter + 1;
     '''
     def __init__(self):
-        whitespace_before_token.__init__(self, 'variable_assignment', '003', lTokens)
-        self.solution = 'Ensure at least a single space exists before the *:=* keyword.'
+        Rule.__init__(self, 'variable_assignment', '003', lTokens)
         self.number_of_spaces = '>=1'
