+++ vsg/rules/whitespace_before_token.py
@@ -9,7 +9,7 @@ from vsg.rules import utils as rules_utils
 from vsg.rules.whitespace_between_tokens import Rule as WhitespaceRule
 
 
-class whitespace_before_token(WhitespaceRule):
+class Rule(WhitespaceRule):
     '''
     Checks for a at least a single space before a token.
 
@@ -35,12 +35,18 @@ class whitespace_before_token(WhitespaceRule):
         lToi = oFile.get_token_and_n_tokens_before_it(self.lTokens, 2)
         for oToi in lToi:
             lTokens = oToi.get_tokens()
-            if isinstance(lTokens[0], parser.carriage_return):
-                continue
-            if isinstance(lTokens[1], parser.carriage_return):
+            if token_is_at_beginning_of_line(lTokens):
                 continue
             if isinstance(lTokens[1], parser.whitespace):
                 lReturn.append(oToi)
             else:
                 lReturn.append(oToi.extract_tokens(1, 2))
         return lReturn
+
+
+def token_is_at_beginning_of_line(lTokens):
+    if isinstance(lTokens[0], parser.carriage_return):
+        return True
+    if isinstance(lTokens[1], parser.carriage_return):
+        return True
+    return False
